types.hpp: remove intrinsic includes
Replace v128 with u128 in some places.
Removed some unused files.

+++ rpcs3/Emu/System.h
@@ -124,7 +124,7 @@ public:
 	std::vector<std::string> argv;
 	std::vector<std::string> envp;
 	std::vector<u8> data;
-	std::vector<u8> klic;
+	std::vector<u128> klic;
 	std::string disc;
 	std::string hdd1;
